# docs/openapi.yaml
openapi: 3.0.3
info:
  title: North Sea Surf School API
  version: "1.0.0"
  description: REST API für Buchung, Verfügbarkeit und Backoffice der Surfschul-Plattform.
servers:
  - url: /api
paths:
...
  /products:
    get:
      summary: Öffentliche Produktliste
      parameters:
        - in: query
          name: locationId
          schema: { type: string }
      responses:
        "200": { description: Produktliste }
...
  /availability:
    get:
      summary: Verfügbare Zeiten und Kapazitäten
      parameters:
        - in: query
          name: productId
          required: true
          schema: { type: string }
        - in: query
          name: locationId
          required: true
          schema: { type: string }
        - in: query
          name: date
          required: true
          schema: { type: string, example: "2026-08-14" }
        - in: query
          name: participants
          schema: { type: integer, default: 1 }
      responses:
        "200":
          description: Slots mit availableSlots, Preis und fehlenden Ressourcen
...
  /reports:
    get:
      summary: Reports inkl. CSV
      parameters:
        - in: query
          name: format
          schema: { type: string, enum: [json, csv] }
...